svggen/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
diffusers==0.24.0
transformers==4.33.2
accelerate==0.23.0
safetensors==0.3.3
pillow==10.0.0
numpy==1.24.4
einops==0.6.1
tqdm==4.66.1
torch==2.0.1
torchvision==0.15.2
pytest==7.4.2
python-json-logger==2.0.7
